Move test utils into test-utils dir
